setup extra test run for selective `gitoxide`-enabled tests
Using a rust config flag we are able to selective run only those
tests that have been cleared to run with `gitoxide`, as part of
a separate test run.

That way we will keep testing all `git2` related code as before
but step by step enable more tests to work correctly with `gitoxide`.

For now it's only planned to run git-related tests, but it's possible
to one-day run the entire test suite if oen were to be willing to pay
the cost.

CI is configured to run separate tests for `gitoxide` as well.